requirements.txt
apns2==0.7.1
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
dj-database-url==0.5.0
dj-email-url==1.0.0
dj-redis-url==0.1.4
Django==2.2.11
django-cors-headers==3.2.1
django-extra-fields==2.0.3
django-filter==2.2.0
django-filters==0.2.1
django-rest-swagger==2.2.0
django-simple-history==2.8.0
django-websocket-redis==0.6.0
djangorestframework==3.11.0
gevent==1.5.0
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
phonenumbers==8.12.1
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.3
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
pywebpush==1.10.2
redis==3.4.1
requests==2.23.0
rsa==4.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
validate-docbr==1.6.1
whitenoise==5.0.1
crypto==1.4.1
